+++ auxiliary/producers/producer/BUILD
@@ -1,7 +1,7 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 
 go_binary(
-    name = "replayer",
+    name = "producer",
     embed = [":go_default_library"],
     visibility = ["//visibility:public"],
 )
@@ -11,10 +11,11 @@ go_library(
     srcs = [
         "server.go"
     ],
-    importpath = "github.com/bptlab/cepta/auxiliary/producers/replayer",
+    importpath = "github.com/bptlab/cepta/auxiliary/producers/producer",
     visibility = ["//visibility:private"],
     deps = [
-        ":producer",
+        ":replayer",
+        "//auxiliary/producers/traindataproducer:livetraindatareplayer",
         "//models/grpc:replayer_go_proto",
         "//osiris/lib:db",
         "//osiris/lib:cli",
@@ -26,19 +27,20 @@ go_library(
         "@com_github_golang_protobuf//ptypes:go_default_library_gen",
         "@com_github_urfave_cli_v2//:go_default_library",
     ],
-)
+) 
 
 go_library(
-    name = "livetraindatareplayer",
+    name = "replayer",
     srcs = [
-        "LiveTrainDataReplayer.go"
+        "Replayer.go"
     ],
     importpath = "github.com/bptlab/cepta/auxiliary/producers/producer/replayer",
-    visibility = ["//visibility:private"],
+    visibility = ["//visibility:public"],
     deps = [
         "//osiris/lib:db",
         "//models/grpc:replayer_go_proto",
         "//osiris/lib:kafkaproducer",
+        "//models/events:live_train_data_go_proto",
         "@com_github_shopify_sarama//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
