+++ auxiliary/producers/producer/server.go
@@ -1,4 +1,4 @@
-package main
+/* package main
 
 import (
 	"context"
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"time"
 
-	livetraindatareplayer "github.com/bptlab/cepta/aux/producers/traindataproducer/livetraindatareplayer"
+	livetraindatareplayer "github.com/bptlab/cepta/auxiliary/producers/producer/livetraindatareplayer"
 	pb "github.com/bptlab/cepta/models/grpc/replayer"
 	libcli "github.com/bptlab/cepta/osiris/lib/cli"
 	libdb "github.com/bptlab/cepta/osiris/lib/db"
@@ -235,3 +235,4 @@ func main() {
 		log.Fatal(err)
 	}
 }
+*/
